requirements.txt
# Core ML/DL
torch
transformers
peft
accelerate
tokenizers # pinned for IndicBERT + Windows

# API
fastapi
uvicorn
pydantic

# UI
streamlit
plotly
matplotlib
seaborn

# Explainability
captum

# Data & Storage
pandas
numpy
pymongo

# News & Web
requests
beautifulsoup4

# Utils
python-dotenv
scikit-learn
langdetect
python-dateutil
tqdm
datasets
fasttext-wheel
shap
huggingface-hub

